pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["JackLi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tzdata = "^2022.7"
coverage = "^7.1.0"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
wand = "^0.6.11"
django = "^4.2.1"
wagtail = "^5.0.1"
wagtailcodeblock = "^1.29.0.0"
django-cors-headers = "^4.2.0"
youtube-dl = {git = "https://github.com/ytdl-org/youtube-dl.git"}


[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
